Title: 앞의 두 수를 더한 수열 2, Time: 137ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240717/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.java
@@ -3,7 +3,7 @@ import java.lang.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int[] dp = new int[91];
+        long[] dp = new long[91];
         dp[1]=1; dp[2]=1;
         int n = sc.nextInt();
         for(int i=3; i<=n; i++){
